feat!: add encodeOrdinary and countTokensOrdinary
The existing encode and countTokens method currently throw an exception if a special
token is encountered. This commit introduces encodeOrdinary which simply encodes a
special token as if it were normal text. This is most likely the behaviour that most
people that are writing an application using the OpenAI API want.

BREAKING CHANGE: Two methods were added to the Encoding interface.