fix: re-add countTokensOrdinary to keep breaking changes to a minimum
The method was deleted during a previous pull request. Technically, it would make
more sense to get rid of the whole differentiation between encode and encodeOrdinary
since there is no practical use-case for consumers of OpenAI models to deal with special
tokens in their LLM input. But to keep backwards compatibility as high as possible, this
commit re-introduces the deleted method and adds benchmarking and tests for it.